src/requirements.txt
fastapi==0.128.0
loguru==0.7.3
python-dotenv==1.2.1
sqlalchemy==2.0.45
uvicorn==0.40.0
asyncpg==0.31.0
httpx==0.28.1
APScheduler==3.11.0
